+++ backend/src/huffman/mod.rs
@@ -1,0 +1,3 @@
+pub mod coding;
+pub mod decoding;
+pub mod models;
