--- backend/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod huffman_coding;
-pub mod models;
